+++ p2p/net/reuseport/dial.go
@@ -1,4 +1,4 @@
-package tcpreuse
+package reuseport
 
 import (
 	"context"
