+++ p2p/net/reuseport/reuseport_plan9.go
@@ -1,4 +1,4 @@
-package tcpreuse
+package reuseport
 
 import (
 	"net"
