+++ p2p/net/reuseport/reuseport_posix.go
@@ -1,7 +1,7 @@
 //go:build !plan9
 // +build !plan9
 
-package tcpreuse
+package reuseport
 
 import (
 	"net"
