+++ p2p/net/reuseport/singledialer.go
@@ -1,4 +1,4 @@
-package tcpreuse
+package reuseport
 
 import (
 	"context"
