+++ p2p/net/reuseport/transport.go
@@ -1,13 +1,13 @@
-// Package tcpreuse provides a basic transport for automatically (and intelligently) reusing TCP ports.
+// Package reuseport provides a basic transport for automatically (and intelligently) reusing TCP ports.
 //
 // To use, construct a new Transport and configure listeners tr.Listen(...).
 // When dialing (tr.Dial(...)), the transport will attempt to reuse the ports it's currently listening on,
 // choosing the best one depending on the destination address.
 //
-// It is recommended to set set SO_LINGER to 0 for all connections, otherwise
+// It is recommended to set SO_LINGER to 0 for all connections, otherwise
 // reusing the port may fail when re-dialing a recently closed connection.
 // See https://hea-www.harvard.edu/~fine/Tech/addrinuse.html for details.
-package tcpreuse
+package reuseport
 
 import (
 	"errors"
