+++ p2p/net/reuseport/transport_test.go
@@ -1,4 +1,4 @@
-package tcpreuse
+package reuseport
 
 import (
 	"net"
